refactor(constants): move cookies constants to `/constants/cookies`
