--- app/api/constants/cookies.ts
@@ -1,1 +1,0 @@
-export const USER_ACCEPT_COOKIES = 'user-accept-cookies'
