+++ app/api/constants/index.ts
@@ -1,3 +1,2 @@
 export * from './api-keys'
-export * from './cookies'
 export * from './errors'
