+++ app/constants/cookies.ts
@@ -1,0 +1,2 @@
+export const USER_ACCEPT_COOKIES = 'user-accept-cookies' as const
+export const USER_ID = 'user-id' as const
