+++ app/constants/index.ts
@@ -1,6 +1,2 @@
-export const USER_ID = 'user-id' as const
-export const ARTIST_ID = 'artist-id' as const
-export const TRACK_ID = 'track-id' as const
-export const ALBUM_ID = 'album-id' as const
-
 export * from './params'
+export * from './cookies'
